feat: Recursive children nodes parsing for ItemCategory
